added linked list functions, removed some null protections

+++ ft_print2dstr.c
@@ -6,15 +6,15 @@
 /*   By: ngontjar <ngontjar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/16 17:22:09 by ngontjar          #+#    #+#             */
-/*   Updated: 2019/10/29 19:56:21 by ngontjar         ###   ########.fr       */
+/*   Updated: 2019/11/22 20:20:30 by ngontjar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_print2dstr(char **array, int count)
+void	ft_print2dstr(char **array, size_t count)
 {
-	int i;
+	size_t	i;
 
 	if (array != NULL)
 	{
