+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: ngontjar <ngontjar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/17 17:33:37 by ngontjar          #+#    #+#             */
-/*   Updated: 2019/10/25 14:25:24 by ngontjar         ###   ########.fr       */
+/*   Updated: 2019/11/22 20:13:30 by ngontjar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *str, int c)
 {
-	int	i;
+	ssize_t	i;
 
 	i = ft_strlen(str);
 	while (i >= 0)
